# Added option to disable full of resources at TR.scala