--- a/src/keys.scala
+++ b/src/keys.scala
@@ -81,6 +81,8 @@ object Keys {
     "TR.scala generating task") in Android
   val typedResources = SettingKey[Boolean]("typed-resources",
     "flag indicating whether to generated TR.scala") in Android
+  val typedResourcesFull = SettingKey[Boolean]("typed-resources-full",
+    "whether full of resources should be generated at TR.scala, default true") in Android
   val typedResourcesIgnores = SettingKey[Seq[String]]("typed-resources-ignores",
     "list of android package names to ignore for TR.scala generation") in Android
   val buildConfigGenerator = TaskKey[Seq[File]]("build-config-generator",