--- a/src/rules.scala
+++ b/src/rules.scala
@@ -599,6 +599,7 @@ object Plugin extends sbt.Plugin with PluginFail {
     proguardScala           <<= autoScalaLibrary,
     retrolambdaEnabled       := false,
     typedResources          <<= autoScalaLibrary,
+    typedResourcesFull       := true,
     typedResourcesIgnores    := Seq.empty,
     typedResourcesGenerator <<= typedResourcesGeneratorTaskDef,
     useProguard             <<= proguardScala,