--- a/src/tasks.scala
+++ b/src/tasks.scala
@@ -299,7 +299,7 @@ object Tasks {
     Resources.generateTR(typedResources.value, rGenerator.value,
       packageForR.value, projectLayout.value, platformApi.value,
       platformJars.value, (scalaVersion in ThisProject).value,
-      libraryProjects.value, typedResourcesIgnores.value, streams.value)
+      libraryProjects.value, typedResourcesFull.value, typedResourcesIgnores.value, streams.value)
   }
 
   def ndkbuild(manager: AndroidSdkHandler, layout: ProjectLayout, args: Seq[String],