Last build of my big project became slower for about 1 min.
So my TR.scala became too fat ~ 3500 LOC (was ~1000), but I don't need for all typed resources.