@pfn 2016-05-12
Thanks, do you have timings for on and off? is scala compilation slower because there's many lines?

in an sbt repl, run a clean compile disabled and not with:

`eval sys.props("sbt.task.timings") = "true"`