@emstlk 2016-05-13
Well I've got some metrics, but I have to notice some changes:
- last time I compared versions 1.5.20 and 1.6.0, but now 1.6.2 and 1.6.3(with my PR)
- my project was separated on few modules and TR became smaller (~2800 LOC)

Build timings with clean project: 

Before
```
[info] dex method count: 154528
[info] Packaged: app-debug-unaligned.apk (31.21MB)
[info] Debug package does not need signing: app-debug-unaligned.apk
[info] zipaligned: app-debug.apk
Total time: 262892.8219 ms
  {file:/Users/emaster/work/android/}app/compile:compileIncremental: 111475.401724 ms
  {file:/Users/emaster/work/android/}app/android:dex: 61120.800881999996 ms
  {file:/Users/emaster/work/android/}app/android:proguard: 28962.861431999998 ms
  {file:/Users/emaster/work/android/}app/android:collectResources: 21693.867999 ms
  {file:/Users/emaster/work/android/}app/android:apkbuild: 8592.863922 ms
  {file:/Users/emaster/work/android/}app/compile:scalastyle: 7511.403085 ms
  {file:/Users/emaster/work/android/}app/android:proguardInputs: 5031.350758 ms
```

and after:
```
[info] dex method count: 152778
[info] Packaged: app-debug-unaligned.apk (31.13MB)
[info] Debug package does not need signing: app-debug-unaligned.apk
[info] zipaligned: app-debug.apk
Total time: 235508.78614699998 ms
  {file:/Users/emaster/work/android/}app/compile:compileIncremental: 94992.77807 ms
  {file:/Users/emaster/work/android/}app/android:dex: 57819.993527 ms
  {file:/Users/emaster/work/android/}app/android:proguard: 29676.801543999998 ms
  {file:/Users/emaster/work/android/}app/android:collectResources: 17752.402498 ms
  {file:/Users/emaster/work/android/}app/android:apkbuild: 8531.224821 ms
  {file:/Users/emaster/work/android/}app/compile:scalastyle: 6545.252786 ms
  {file:/Users/emaster/work/android/}app/android:proguardInputs: 4866.260839 ms
  {file:/Users/emaster/work/android/}app/compile:compile: 4089.3132109999997 ms
```

Also I made builds with xml changes:

Before:
```
[info] dex method count: 189940
[info] Packaged: app-debug-unaligned.apk (33.49MB)
[info] Debug package does not need signing: app-debug-unaligned.apk
[info] zipaligned: app-debug.apk
Total time: 187440.615955 ms
  {file:/Users/emaster/work/android/}app/compile:compileIncremental: 94305.923788 ms
  {file:/Users/emaster/work/android/}app/android:dex: 65883.017712 ms
  {file:/Users/emaster/work/android/}app/android:apkbuild: 8277.476741 ms
  {file:/Users/emaster/work/android/}app/compile:scalastyle: 4830.728107 ms
  {file:/Users/emaster/work/android/}app/compile:compile: 3480.4259039999997 ms
  {file:/Users/emaster/work/android/}app/android:packageResources: 2592.0621739999997 ms
  {file:/Users/emaster/work/android/}app/android:proguardInputs: 1668.354337 ms
  {file:/Users/emaster/work/android/}app/android:rGenerator: 1524.781963 ms
```

After:
```
[info] dex method count: 188156
[info] Packaged: app-debug-unaligned.apk (33.36MB)
[info] Debug package does not need signing: app-debug-unaligned.apk
[info] zipaligned: app-debug.apk
Total time: 170504.086221 ms
  {file:/Users/emaster/work/android/}app/compile:compileIncremental: 76630.050978 ms
  {file:/Users/emaster/work/android/}app/android:dex: 67165.399379 ms
  {file:/Users/emaster/work/android/}app/android:apkbuild: 8129.305799999999 ms
  {file:/Users/emaster/work/android/}app/compile:scalastyle: 4784.257560999999 ms
  {file:/Users/emaster/work/android/}app/compile:compile: 3659.26592 ms
  {file:/Users/emaster/work/android/}app/android:packageResources: 2664.404465 ms
  {file:/Users/emaster/work/android/}app/android:rGenerator: 1406.671464 ms
```